More work on dictionary
